Changed layout parameters to be device dependent, should look find on any device now.
Reset prevPos tracker on motion event action_up, which was causing a bug where if you pressed just one letter, and then stopped, it would clear the letter, but the next time it wouldn't, because prevPos wasn't being cleared.

There's also a bug where any single letter is evaluated as true in the trie. Keeping it in for now since it helps with debugging when I get a set of letters I don't like, but needs to be fixed before release, as most single letters aren't words.